feature/integration-user-screen: separei a mutation em um hook conforme comentário
